requirements.txt
paramiko
python-dotenv
EbookLib
beautifulsoup4
